Logic explanation
Setting up a tournament array which computes the maximum of all elements entered in the array
say there are 8 elements
you have to insert the elements in an array starting with array index N and ending with 2N-1 (e.g. if N is 8, fill 8 elements array between 8 and 15)
compare elements that are between array indices N and 2N-1 in pairwise format
e.g. first comparison between array element 2N-1 and 2N-2 (if N = 8 it is between elements with array indices 14 and 15)
second comparison between array element 2N-3 and 2N-4 (if N = 8 it is between elements with array indices 12 and 13) and so on. 
place the maximum of first comparison in the array with index counting down starting from 7 (i.e. 14/2 i.e. 2N-2/2) 
continue this as long as you don't reach the array index  1 
the output in the array index 1 is the maximum of the N numbers of the array